fixed adding of rules to collaboration to take place on the entity so that the rule is cascaded
