+++ src/domain/collaboration/collaboration/collaboration.module.ts
@@ -15,6 +15,7 @@ import { CanvasModule } from '@domain/common/canvas/canvas.module';
 import { AspectModule } from '../aspect/aspect.module';
 import { ActivityAdapterModule } from '@services/adapters/activity-adapter/activity.adapter.module';
 import { NotificationAdapterModule } from '@services/adapters/notification-adapter/notification.adapter.module';
+import { CommunityPolicyModule } from '@domain/community/community-policy/community.policy.module';
 
 @Module({
   imports: [
@@ -23,6 +24,7 @@ import { NotificationAdapterModule } from '@services/adapters/notification-adapt
     AuthorizationPolicyModule,
     AuthorizationModule,
     CalloutModule,
+    CommunityPolicyModule,
     NamingModule,
     RelationModule,
     CanvasModule,
